+++ src/Point/Point.h
@@ -35,7 +35,7 @@ public:
   static unsigned int identifierIndex ();
 
   /// Get method for ordinal.
-  int ordinal () const;
+  double ordinal () const;
 
   /// Accessor for graph position.
   QPointF posGraph () const;
@@ -47,7 +47,7 @@ public:
   static void setIdentifierIndex (unsigned int identifierIndex);
 
   /// Set the ordinal used for ordering Points.
-  void setOrdinal (int ordinal);
+  void setOrdinal (double ordinal);
 
   /// Set method for position in graph coordinates.
   void setPosGraph (const QPointF &posGraph);
@@ -68,7 +68,7 @@ private:
   QString m_identifier;
   QPointF m_posScreen;
   QPointF m_posGraph;
-  int m_ordinal;
+  double m_ordinal;
 
   static unsigned int m_identifierIndex; // For generating unique identifiers
 };
